move engine tests to a separate file. they were getting too long.
